feat: add styles, refactor code and complete task
