+++ tasks/dayNigth/src/index.js
@@ -1,7 +1,7 @@
 
 import "core-js/stable";
 import "regenerator-runtime/runtime";
-import App from "./app";
+import App from "./App";
 
 const app = new App();
