+++ tasks/dayNigth/src/styles/index.scss
@@ -7,96 +7,129 @@ html {
 body {
   max-width: 800px;
   margin: 0 auto;
+  color: var(--fifth-color);
   background-color: var(--secondary-color);
 }
-
 
 .header {
   display: flex;
   justify-content: space-between;
+  background-color: var(--secondary-color);
 
-  .switch input {
-    position: absolute;
-    opacity: 0;
-  }
+  .theme-switch-container {
 
-  .switch {
-    display: inline-block;
-    font-size: 20px;
-    height: 1em;
-    width: 2em;
-    background: #BDB9A6;
-    border-radius: 1em;
-
-    &__icon {
-      height: 1em;
-      width: 1em;
-      border-radius: 1em;
-      background: #FFF;
-      box-shadow: 0 0.1em 0.3em rgba(0, 0, 0, 0.3);
-      -webkit-transition: all 300ms;
-      -moz-transition: all 300ms;
-      transition: all 300ms;
+    .switch input {
+      position: absolute;
+      opacity: 0;
     }
 
-    input:checked+.switch__icon {
-      -webkit-transform: translate3d(100%, 0, 0);
-      -moz-transform: translate3d(100%, 0, 0);
-      transform: translate3d(100%, 0, 0);
-      background: #000;
+    .switch {
+      height: 1em;
+      width: 2em;
+      margin: 10px;
+      display: inline-block;
+      font-size: 20px;
+      background-color: var(--third-color);
+      border-radius: 1em;
+
+      &__icon {
+        height: 1em;
+        width: 1em;
+        border-radius: 1em;
+        background-color: var(--six-color);
+        box-shadow: 0 0.1em 0.3em var(--fifth-color);
+        -webkit-transition: all 300ms;
+        -moz-transition: all 300ms;
+        transition: all 300ms;
+        cursor: pointer;
+      }
+
+      input:checked+.switch__icon {
+        -webkit-transform: translate3d(100%, 0, 0);
+        -moz-transform: translate3d(100%, 0, 0);
+        transform: translate3d(100%, 0, 0);
+        background: #000;
+      }
+    }
+  }
+
+  .login-container {
+    margin: 10px;
+
+    &__title {
+      padding: 5px;
+      color: var(--fifth-color);
+    }
+
+    &__input-container {
+      padding: 5px;
+
+      &__input {
+        padding: 5px;
+      }
+    }
+  }
+
+  .submit-container {
+    margin: 10px;
+
+    &__title {
+      padding: 5px;
+      color: var(--fifth-color);
     }
   }
 }
-
-
-
-
-
-
 
 .wrapper {
   margin: 0 auto;
   display: flex;
+  flex-direction: column;
+  background-color: var(--six-color);
 }
 
-.dropdown {
-  min-width: 220px;
-  margin: 10px;
-  position: relative;
-  display: block;
-  background-color: #acacac;
-  color: #fff;
+.dropdowns {
+  display: flex;
 
-  &:hover .dropdown__content {
+  .dropdown {
+    min-width: 220px;
+    margin: 10px;
+    position: relative;
     display: block;
-  }
+    background-color: var(--third-color);
+    color: var(--first-color);
 
-  &__title {
-    padding: 10px;
-    font-size: 24px;
-    font-style: italic;
-    font-weight: bold;
-  }
+    &:hover .dropdown__content {
+      display: block;
+    }
 
-  &__content {
-    display: none;
-    position: absolute;
-    background-color: #f9f9f9;
-    width: 220px;
-    box-shadow: 0px 8px 16px 0px rgba(0, 0, 0, 0.2);
-    padding: 12px 16px;
-    z-index: 1;
-    max-height: 400px;
-    overflow: auto;
+    &__title {
+      padding: 10px;
+      font-size: 24px;
+      font-style: italic;
+      font-weight: bold;
+    }
 
-    &__link {
-      padding: 0 5px;
-      cursor: pointer;
-      color: #000;
+    &__content {
+      display: none;
+      position: absolute;
+      background-color: var(--first-color);
+      width: 220px;
+      box-shadow: 0px 8px 16px 0px var(--fifth-color);
+      padding: 12px 16px;
+      z-index: 1;
+      max-height: 400px;
+      overflow: auto;
+      background-color: var(--secondary-color);
 
-      &:hover {
-        background-color: #acacac;
-        color: #fff;
+      &__link {
+        padding: 0 5px;
+        cursor: pointer;
+        color: var(--fifth-color);
+
+        &:hover {
+          background-color: var(--third-color);
+
+        }
       }
     }
   }
@@ -114,9 +147,9 @@ body {
 .api-info {
   margin: 10px 30px;
   padding: 10px;
-  border: 1px solid #acacac;
+  border: 1px solid var(--third-color);
   border-radius: 10px;
-  background-color: #ebebeb;
+  background-color: var(--secondary-color);
 
   &__title {
     font-size: 18px;
@@ -127,10 +160,30 @@ body {
   }
 
   &__link {
-    color: #000;
+    color: var(--fifth-color);
   }
 }
 
 .title-error {
   font-size: 40px;
+}
+
+button {
+  margin: 5px;
+  font-size: 16px;
+  font-weight: bold;
+  padding: 5px 10px;
+  border-radius: 5px;
+  border: 1px solid var(--fifth-color);
+  color: var(--fifth-color);
+  background-color: var(--six-color);
+  cursor: pointer;
+
+  &:hover {
+    color: var(--secondary-color);
+    background-color: var(--fifth-color);
+  }
+
+  &:active {}
+
 }
